+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.0.1
+-----
+
+* Switching to use leafo's caching by using server() instead of compile().
+* Requires a writable directory and depends on Leafo/scssphp ^0.1.1 now.
 
 1.0.0
 -----